// client/src/api/model/costType.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

export type CostType = typeof CostType[keyof typeof CostType];


export const CostType = {
  None: 'None',
  Strain: 'Strain',
  Story_Point: 'Story Point',
} as const;