// client/src/api/model/statsType.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

export type StatsType = (typeof StatsType)[keyof typeof StatsType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const StatsType = {
  Wounds: "Wounds",
  Strain: "Strain",
} as const;